1. ui fixes

+++ app/src/main/java/com/onelink/nrlp/android/utils/ValidationUtils.kt
@@ -130,7 +130,7 @@ object ValidationUtils {
     }
 
     fun isAmountValid(string: String): Boolean {
-        return string.isNotEmpty() && string.toInt() > 0
+        return string.isNotEmpty() && string.toFloat() > 0
     }
 
     fun isTransactionNoValid(string: String): Boolean {
